pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gantrygraph"
version = "0.1.0"
description = "Autonomous AI agent framework — visual computer use, MCP tools, and multi-agent swarms"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "GantryGraph Contributors", email = "hello@gantrygraph.com" },
]
keywords = [
    "ai-agents", "autonomous-agents", "computer-use", "langgraph",
    "mcp", "model-context-protocol", "desktop-automation", "browser-automation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=2.0.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "mss>=9.0.0",
    "anyio>=4.0",
    "typing-extensions>=4.9.0",
]

[project.urls]
Homepage = "https://gantrygraph.com"
Documentation = "https://gantrygraph.com"
Repository = "https://github.com/GantryGraph/GantryGraph"
"Bug Tracker" = "https://github.com/GantryGraph/GantryGraph/issues"
Changelog = "https://github.com/GantryGraph/GantryGraph/releases"

[project.optional-dependencies]
desktop   = ["pyautogui>=0.9.54"]
browser   = ["playwright>=1.40.0"]
cloud     = ["fastapi>=0.100.0", "uvicorn>=0.23.0"]
memory    = ["chromadb>=0.5.0"]
telemetry = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]
all = ["gantrygraph[desktop,browser,cloud,memory,telemetry]"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "langchain-anthropic>=0.3.0",
    "httpx>=0.27.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gantrygraph"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
strict = true
python_version = "3.11"
mypy_path = "src"
explicit_package_bases = true

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
